Restructure configuration module to provide more flexibility
