+++ dementor/protocols/ldap.py
@@ -36,7 +36,9 @@ from impacket.ldap.ldapasn1 import (
 )
 from pyasn1.codec.ber import encoder as BEREncoder, decoder as BERDecoder
 
-from dementor.config import SessionConfig, TomlConfig, Attribute as A, get_value
+from dementor.config.toml import TomlConfig, Attribute as A
+from dementor.config.session import SessionConfig
+from dementor.config.util import get_value
 from dementor.logger import ProtocolLogger
 from dementor.servers import (
     ThreadingTCPServer,
